Refactor GeonotebookStack to use OrderedDict
This also seperates the ideas of a user layer (now in
GeonotebookStack._layers) and a system layer (stored in
GeonotebookStack._system_layers).  System layers are for things like the
base map,  annotations, etc -  things the user should probably not be
adding, deleting, etc.  GeonotebookLayers may hint that they are
system_layers through anattribute called _system_layer.

Additionally this adds a GeonotebookLayer attribute _expose_as. If set
to a string the layer will be exposed as an attribute on the
GeonotebookStack. Any layer may be exposed, but the intention of the
feature is to allow exposing system layers (e.g.  M.layers.annotation is
made available because the annotation layer is added with a kwarg
'expose_as=annotation').