Extend protocol negotiation and RPC calls
This extends the protocol negotiation on the javascript side to maintain
state information about whether or not function arguments are 'optional'
or not (based on whether they have default arguments). Additionally the
'params' section of of python to JS RPC calls has been extended so that
it is no longer a list of values,  but a list of dicts that maintain
key/value pairs.  This allows us to correctly resolve params sent to
javascript with javascript function arguments.