Add utilities for sending/receiving logs from external applications
This will allow any application that can talk to the notebook over HTTP
to POST log records as JSON to the /log endpoint and have them be processed and
logged within the notebook application.

It also adds the "client" logging handler which can talk to this
endpoint. JsonHTTPHandler is a slight modification of the HTTPHandler,
due to the fact that we would rather use JSON as the messaging medium
rather than url encoded messages in order to avoid casting values as
integers, floats, etc.