Remove hashing logic related to KTile provider
This fixes an issue where rendering layers was occasionally throwing an
error such as:
RuntimeError:
`/tmp/geonb4Y4vo3/L57.Globe.month09.2010.hh09vv04.h6v1.doy247to273.NBAR.v3.0_504169755319211990.vrt'
not recognised as a supported file format.

The .vrt file was size 0 (not written to disk yet) by the time the
javascript was trying to render it. This was due to an error in the
hashing causing generate_vrt to be called.

Hashing has been moved into style object and layers are named after
their data+style.